Address Issue #441
It looks like there has been a change in the behaviour of the tab
container at 2.5. Tabs within the container will always fill 100% of the
width of the container, in previous releases there was a hidden tab that
was always added that allowed you to control the width of the other tabs
and this hidden tab would fill the gap to make it up to 100%.

It looks like this has now been removed, so the fix was to manually add
this hidden tab, a few CSS changes where also required to make the UI
correct.